requirements.txt
# 基础 HTTP 请求
requests>=2.31.0

# 若想打印或分析资产数据（可选）
pandas>=2.1.0
tabulate>=0.9.0
python-okx>=0.4.0
openai==1.43.1
python-dotenv==1.0.1
pandas==2.2.2
ta==0.11.0